requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
crewai==0.157.0
crewai-tools==0.60.0
google-generativeai==0.8.3
google-cloud-firestore==2.18.0
firebase-admin==6.5.0
pydantic==2.8.2
pydantic-settings==2.5.2
python-multipart==0.0.6
asyncio==3.4.3
python-dotenv==1.0.0
structlog==23.2.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
pandas==2.1.3
openpyxl==3.1.2
